+++ config/influence.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * Mai Engine.
+ *
+ * @package   BizBudding\MaiEngine
+ * @link      https://bizbudding.com
+ * @author    BizBudding
+ * @copyright Copyright © 2020 BizBudding
+ * @license   GPL-2.0-or-later
+ */
+
+return [
+	'demos' => [
+		// 'travel'      => 0,
+		// 'home-garden' => 0,
+	],
+	'global-styles' => [
+		'colors' => [
+			'link'      => '#a0d9ca',
+			'primary'   => '#a0d9ca',
+			'secondary' => '#173c6e',
+			'heading'   => '#333333',
+			'body'      => '#333333',
+			'alt'       => '#f9f9f9',
+			// 'black'     => '#222222',
+		],
+		'fonts' => [
+			'body'       => 'Karla:400',
+			'heading'    => 'Montserrat:400',
+			// 'subheading' => 'Playfair Display:500italic',
+		],
+	],
+	'image-sizes' => [
+		'add' => [
+			'landscape' => '3:2',
+			'portrait'  => '2:3',
+		],
+	],
+	'settings' => [],
+	'custom-functions' => [],
+];