+++ lib/functions/plugins.php
@@ -104,6 +104,66 @@ function mai_woocommerce_breakpoint() {
 	return mai_get_unit_value( mai_get_breakpoint( $breakpoint ), 'px' );
 }
 
+add_filter( 'woocommerce_product_loop_start', 'mai_product_loop_start_columns' );
+/**
+ * Adds column count as inline custom properties.
+ *
+ * @since 1/4/21
+ *
+ * @param array $html The existing loop start HTML.
+ *
+ * @return string
+ */
+function mai_product_loop_start_columns( $html ) {
+
+	$count = wc_get_loop_prop( 'columns' );
+
+	if ( ! is_numeric( $count ) ) {
+		return $html;
+	}
+
+	$dom   = mai_get_dom_document( $html );
+	$first = mai_get_dom_first_child( $dom );
+
+	if ( ! $first ) {
+		return $html;
+	}
+
+	// Get the columns breakpoint array.
+	$columns = mai_get_breakpoint_columns(
+		[
+			'columns' => $count,
+		]
+	);
+
+	$style  = $first->getAttribute( 'style' );
+	$style .= sprintf( '--columns-xs:%s;', $columns['xs'] );
+	$style .= sprintf( '--columns-sm:%s;', $columns['sm'] );
+	$style .= sprintf( '--columns-md:%s;', $columns['md'] );
+	$style .= sprintf( '--columns-lg:%s;', $columns['lg'] );
+
+	$first->setAttribute( 'style', $style );
+
+	return str_replace( '</ul>', '', $dom->saveHTML() );
+}
+
+add_filter( 'woocommerce_pagination_args', 'mai_woocommerce_pagination_previous_next_text' );
+/**
+ * Changes the adjacent entry previous and next link text.
+ *
+ * @since 1/4/21
+ *
+ * @param array $args The pagination args.
+ *
+ * @return array
+ */
+function mai_woocommerce_pagination_previous_next_text( $args ) {
+	$args['prev_text'] = esc_html__( '← Previous', 'mai-engine' );
+	$args['next_text'] = esc_html__( 'Next →', 'mai-engine' );
+
+	return $args;
+}
+
 /**
  * Filter single product post_class.
  * Make sure it's only run on the main product entry wrap.
@@ -142,7 +202,7 @@ function mai_woocommerce_product_single_class( $classes ) {
  *
  * @return bool
  */
-add_filter( 'woocommerce_price_trim_zeros', '__return_true' );
+add_filter( 'woocommerce_price_trim_zeros', '__return_true', 8 );
 
 add_filter( 'woocommerce_cart_item_remove_link', 'mai_woocommerce_cart_item_remove_icon' );
 /**